requirements-ci.txt
# CI Dependencies — Single Source of Truth
# All workflows install from this file.
# Update versions here ONLY — do not pin in individual workflow files.
#
# Install: pip install -r requirements-ci.txt
# Dev:     pip install -r requirements-ci.txt -r requirements-dev.txt
#
# Sibling parity: Cognitive.Engine.Graphs/requirements-ci.txt

# ── Runtime dependencies ─────────────────────────────────────────────────────
aiofiles>=23.0.0
asyncpg
constellation-node-sdk @ git+https://github.com/cryptoxdog/Gate_SDK.git@main
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
pyyaml>=6.0
redis>=5.0.0
sqlalchemy
structlog>=24.0.0
perplexityai>=0.2.0
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-exporter-otlp-proto-grpc>=1.27.0
opentelemetry-instrumentation-fastapi>=0.48b0
opentelemetry-instrumentation-httpx>=0.48b0
opentelemetry-instrumentation-redis>=0.48b0

# ── Lint & type checking ─────────────────────────────────────────────────────
ruff>=0.4.0,<1.0.0
mypy>=1.10.0,<2.0.0

# ── Type stubs ───────────────────────────────────────────────────────────────
types-PyYAML>=6.0.12
types-redis>=4.6.0

# ── Testing ──────────────────────────────────────────────────────────────────
pytest==9.0.2
pytest-asyncio>=1.0.0,<2.0.0
pytest-cov==7.1.0
pytest-xdist>=3.0.0
pytest-timeout>=2.0.0
pytest-httpx>=0.34.0
respx>=0.22.0
coverage==7.13.5

# ── Security scanning ────────────────────────────────────────────────────────
bandit>=1.9.0
pip-audit>=2.9.0
safety>=3.0.0

# ── License compliance ───────────────────────────────────────────────────────
pip-licenses>=4.0.0
